mcp-servers/data-platform/requirements.txt
# MCP SDK
mcp>=0.9.0

# Data Processing
pandas>=2.0.0
pyarrow>=14.0.0

# PostgreSQL/PostGIS
asyncpg>=0.29.0
geoalchemy2>=0.14.0
shapely>=2.0.0

# dbt
dbt-core>=1.9.0
dbt-postgres>=1.9.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.23.0
